refactor(scripts): #72 2.2-5 — extract BLS/BEA/Census planner+runner blocks
Move three US timeseries-API source blocks to dedicated modules:

  scripts/validate/bls.py    232 LOC
  scripts/validate/bea.py    274 LOC
  scripts/validate/census.py 209 LOC

Each module owns its constants (`*_OBS_EXPECTED_FIELDS`,
`_BLS_PERIOD_RE`), source-specific dataclass (`BLSProbe`,
`BEAProbe`, `CensusProbe`), planner, diff helper, parser dry-run,
and runner. The validator drops 3,406 → 2,796 LOC (−610).

Census previously called BLS's `_bls_probe_token` (cross-source
import of a 5-line slugify helper) — each per-source module now has
its own private slugify, removing that coupling. Three identical
copies is simpler than promoting a trivial helper to `_shared.py`.

Test-compat note: 9 scaffold tests reach into the validator module
via `spec_from_file_location` and access private helpers directly
(`validator._diff_bls_observation`, `validator._try_parse_bea`,
`validator._row_matches_census_probe`, etc.). The validator's
import block re-exports those private names so the existing test
loaders keep finding them — same pattern that slice 2.2-9 will
formalize as the final shim.

Tests: 9 scaffold tests at 135 passed; broader subset at 695
passed. Manual `--provider {bls,bea,census}` dry-runs all plan
correctly.

Refs #72.
