+++ scripts/validate_calendar_acquisition.py
@@ -77,27 +77,39 @@ from scripts.validate.eodhd import (  # noqa: E402
     run_eodhd_probe,
     try_parse_eodhd,
 )
+from scripts.validate.bls import (  # noqa: E402
+    BLSClient,
+    BLSProbe,
+    BLS_INDICATOR_REGISTRY,
+    BLS_OBS_EXPECTED_FIELDS,
+    _diff_bls_observation,
+    _try_parse_bls,
+    plan_bls_probes,
+    run_bls_probe,
+)
+from scripts.validate.bea import (  # noqa: E402
+    BEAClient,
+    BEAProbe,
+    BEA_INDICATOR_REGISTRY,
+    BEA_OBS_EXPECTED_FIELDS,
+    _diff_bea_observation,
+    _try_parse_bea,
+    plan_bea_probes,
+    run_bea_probe,
+)
+from scripts.validate.census import (  # noqa: E402
+    CENSUS_EITS_EXPECTED_FIELDS,
+    CENSUS_INDICATOR_REGISTRY,
+    CensusEITSClient,
+    CensusProbe,
+    _diff_census_row,
+    _row_matches_census_probe,
+    _try_parse_census,
+    plan_census_probes,
+    run_census_probe,
+)
 
 # Scaffold under test — we import and exercise exactly what production uses.
-from ingestion.calendar.bls_api import (  # noqa: E402
-    INDICATOR_REGISTRY as BLS_INDICATOR_REGISTRY,
-    parse_observation as parse_bls_observation,
-)
-from ingestion.timeseries.scrapers.bls import BLSClient  # noqa: E402
-
-from ingestion.calendar.bea_api import (  # noqa: E402
-    INDICATOR_REGISTRY as BEA_INDICATOR_REGISTRY,
-    parse_observation as parse_bea_observation,
-)
-from ingestion.timeseries.scrapers.bea import BEAClient  # noqa: E402
-
-from ingestion.calendar.census_api import (  # noqa: E402
-    CensusEITSClient,
-    CensusEITSObservation,
-    INDICATOR_REGISTRY as CENSUS_INDICATOR_REGISTRY,
-    parse_observation as parse_census_observation,
-)
-
 from ingestion.calendar.ism_api import (  # noqa: E402
     ISM_RELEASE_CALENDAR_URL,
     ISM_REPORTS_URL,
@@ -389,628 +401,6 @@ def _fmt_field_list(fields: list[str]) -> str:
         return "(none)"
     return ", ".join(f"`{f}`" for f in fields)
 
-
-# ──────────────────────────────────────────────────────────────────────────
-# BLS probe plan + runner
-# ──────────────────────────────────────────────────────────────────────────
-
-# Field set each BLSObservation carries in its ``raw`` dict — this is what
-# the BLS API returns per observation row inside ``Results.series[].data[]``.
-# The parser reads ``year`` + ``period`` to compute ``date``, ``value`` for
-# the numeric, and keeps the entire dict in ``BLSObservation.raw`` so
-# footnote-only revisions register a new content hash on the calendar side.
-BLS_OBS_EXPECTED_FIELDS: frozenset[str] = frozenset({
-    "year", "period", "periodName", "value", "footnotes", "latest",
-})
-
-# Structural assertions on a parsed BLSObservation:
-#   - period matches M01..M13 (monthly) / Q01..Q05 (quarterly) / A01 / S01..S02.
-#   - value is numeric (the client's _parse_observation already coerces).
-#   - date is ISO YYYY-MM-DD.
-_BLS_PERIOD_RE = re.compile(r"^(M0[1-9]|M1[0-3]|Q0[1-5]|A01|S0[1-2])$")
-
-
-@dataclass
-class BLSProbe:
-    """One BLS live-validation probe — a single series + year window.
-
-    Separate dataclass from :class:`Probe` because BLS probes drive
-    ``BLSClient.get_series_single`` rather than a raw-path HTTP call;
-    the generic Probe shape (``path``, ``params``, ``rows_key``,
-    ``subtype``) doesn't fit cleanly.
-    """
-
-    name: str
-    series_id: str
-    indicator: str          # canonical token (``"CPI"``, ``"NFP"``)
-    description: str
-    start_year: int
-    end_year: int
-
-
-def plan_bls_probes() -> list[BLSProbe]:
-    """One probe per entry in the BLS calendar INDICATOR_REGISTRY.
-
-    P1c expanded the whitelist from 2 anchors (CPI + NFP) to 11
-    indicators covering the full BLS headline set. Each probe hits
-    one series with a two-year window — the current year plus the
-    prior year — which guarantees coverage across the monthly /
-    quarterly release cadences the whitelist mixes. Total BLS API
-    usage per live run is one request per series (~11), well under
-    the 500-requests-per-day budget.
-    """
-    now_year = datetime.now(timezone.utc).year
-    probes: list[BLSProbe] = []
-    for series_id, spec in BLS_INDICATOR_REGISTRY.items():
-        token = _bls_probe_token(spec.indicator)
-        probes.append(
-            BLSProbe(
-                name=f"{token}_two_year_window",
-                series_id=series_id,
-                indicator=token.upper(),
-                description=spec.title,
-                start_year=now_year - 1,
-                end_year=now_year,
-            )
-        )
-    return probes
-
-
-def _bls_probe_token(indicator_label: str) -> str:
-    """Slugify an indicator label into a stable probe-name token."""
-    token = indicator_label.strip().lower()
-    token = re.sub(r"[^a-z0-9]+", "_", token).strip("_")
-    return token or "indicator"
-
-
-def _diff_bls_observation(raw_row: dict[str, Any]) -> RowDiff:
-    """Field diff for one BLS observation row (the ``raw`` dict)."""
-    observed = set(raw_row.keys())
-    diff = RowDiff(
-        observed_fields=sorted(observed),
-        read_by_parser=sorted(observed & BLS_OBS_EXPECTED_FIELDS),
-        ignored_by_parser=sorted(
-            (BLS_OBS_EXPECTED_FIELDS & observed) - BLS_OBS_EXPECTED_FIELDS
-        ),  # empty by construction — kept for shape symmetry
-        unknown_observed=sorted(observed - BLS_OBS_EXPECTED_FIELDS),
-        missing_expected=sorted(BLS_OBS_EXPECTED_FIELDS - observed),
-    )
-
-    period = raw_row.get("period", "")
-    if period and not _BLS_PERIOD_RE.match(str(period)):
-        diff.type_warnings.append(
-            f"period={period!r} doesn't match M01-M13 / Q01-Q05 / A01 / "
-            f"S01-S02 shape"
-        )
-    value = raw_row.get("value")
-    if value is None:
-        diff.type_warnings.append("value missing — parser would drop this row")
-    elif not isinstance(value, str):
-        diff.type_warnings.append(
-            f"value is {type(value).__name__}={value!r} — parser expects str"
-        )
-    return diff
-
-
-def _try_parse_bls(obs) -> tuple[bool, str]:
-    """Dry-run the calendar-side BLS projector on one observation."""
-    spec = BLS_INDICATOR_REGISTRY.get(obs.series_id)
-    if spec is None:
-        return False, f"no BLS calendar spec for series_id={obs.series_id!r}"
-    try:
-        raw, event = parse_bls_observation(
-            obs, snapshot_epoch_ms=1_700_000_000_000, spec=spec,
-        )
-        return True, (
-            f"ok indicator={spec.indicator} event_id={raw.provider_event_id[:10]}…"
-        )
-    except Exception as exc:
-        return False, f"{type(exc).__name__}: {exc}"
-
-
-def run_bls_probe(client: BLSClient, probe: BLSProbe) -> ProbeResult:
-    """Execute one BLS probe and return a populated :class:`ProbeResult`.
-
-    Runs via ``BLSClient.get_series_single`` — reuses the production
-    transport layer so the probe exercises exactly the path a recurring
-    fetch would. Rate-limit / auth errors surface as ``http_error``;
-    empty responses surface as ``ok`` with ``row_count=0`` so the
-    report still prints the probe card.
-    """
-    # BLS uses a POST body rather than a GET path; the audit trail
-    # needs to describe both the wire shape (endpoint) and the
-    # body-level identity (series id + year window) so a reader can
-    # reproduce the call from the report alone.
-    generic = Probe(
-        name=probe.name,
-        path="POST https://api.bls.gov/publicAPI/v2/timeseries/data/",
-        description=probe.description,
-        expected_shape="list[BLSObservation] after client parse",
-        expected_fields=BLS_OBS_EXPECTED_FIELDS,
-    )
-    result = ProbeResult(probe=generic, status="skipped")
-    result.request_path = (
-        f"{generic.path} "
-        f"seriesid=[{probe.series_id}] "
-        f"startyear={probe.start_year} endyear={probe.end_year}"
-    )
-
-    if not client.api_key:
-        result.status = "auth_missing"
-        result.notes.append("BLS_API_KEY not set — probe skipped")
-        return result
-
-    import time as _time
-    t0 = _time.monotonic()
-    try:
-        observations = client.get_series_single(
-            probe.series_id,
-            start_year=probe.start_year,
-            end_year=probe.end_year,
-        )
-    except Exception as exc:
-        result.status = "http_error"
-        result.notes.append(f"{type(exc).__name__}: {exc}")
-        return result
-    result.http_elapsed_ms = (_time.monotonic() - t0) * 1000
-
-    result.status = "ok"
-    result.row_count = len(observations)
-    if not observations:
-        result.notes.append("BLS returned zero observations for this window")
-        return result
-
-    # Sort newest-first so the sample row + diff reflect the most recent
-    # release — that's the one whose shape we care about for schedule
-    # adherence audits.
-    observations = sorted(observations, key=lambda o: o.date, reverse=True)
-    sample = observations[0]
-    result.sample_row = sample.raw or {
-        "series_id": sample.series_id,
-        "date": sample.date,
-        "value": sample.value,
-        "period": sample.period,
-    }
-    if sample.raw:
-        result.field_diff = _diff_bls_observation(sample.raw)
-
-    # period / periodName tallies — surfaces mixed-frequency responses
-    # (monthly + annual average rows) when the client passes
-    # annual_average=False but BLS attaches A01 anyway.
-    period_counter: Counter = Counter()
-    for obs in observations:
-        period_counter[repr(obs.period)] += 1
-    result.enum_counters = {"period": period_counter}
-
-    sample_n = min(10, len(observations))
-    result.parse_attempts = sample_n
-    for obs in observations[:sample_n]:
-        ok, msg = _try_parse_bls(obs)
-        if ok:
-            result.parse_successes += 1
-        else:
-            if len(result.parse_error_samples) < 3:
-                result.parse_error_samples.append(msg)
-    return result
-
-
-# ──────────────────────────────────────────────────────────────────────────
-# BEA probe plan + runner
-# ──────────────────────────────────────────────────────────────────────────
-
-# Fields each BEAObservation carries in its ``raw`` dict — keys the parser
-# hashes on (``_HASH_FIELDS = (DataValue, LineDescription, NoteRef)``) plus
-# the identity fields (``TimePeriod``, ``LineNumber``). Anything observed
-# outside this set surfaces as UNKNOWN_OBSERVED — the prompt to decide
-# whether a new upstream column should be read or explicitly ignored.
-BEA_OBS_EXPECTED_FIELDS: frozenset[str] = frozenset({
-    "TimePeriod", "DataValue", "LineNumber", "LineDescription", "NoteRef",
-})
-
-
-@dataclass
-class BEAProbe:
-    """One BEA live-validation probe — a single series + year window.
-
-    Mirrors :class:`BLSProbe`. BEA's query surface differs from BLS
-    (``GET /api/data`` with ``DatasetName`` / ``TableName`` / ``Frequency``
-    / ``Year`` rather than a POST body), so this dataclass carries the
-    call coordinates the runner needs; the generic :class:`Probe` shape
-    still carries the report metadata on the :class:`ProbeResult`.
-    """
-
-    name: str
-    series_id: str
-    indicator: str          # canonical token (``"GDP"``, ``"PERSONAL_INCOME"``)
-    dataset: str            # BEA dataset name (``"NIPA"``, ``"ITA"``)
-    table: str              # BEA table code (``"T10101"``, ``"T20600"``)
-    line_number: str        # BEA line within the table (``"1"``)
-    frequency: str          # BEA frequency code (``"Q"``, ``"M"``, ``"A"``)
-    description: str
-    start_year: int
-    end_year: int
-
-
-def plan_bea_probes() -> list[BEAProbe]:
-    """One probe per entry in the BEA calendar ``INDICATOR_REGISTRY``.
-
-    Each probe hits one ``(dataset, table, frequency)`` coordinate with
-    a two-year window — current year plus prior — which guarantees
-    coverage across the monthly + quarterly cadences the whitelist
-    mixes. The runner filters the returned rows down to the probe's
-    ``line_number`` after the response lands (BEA returns every line
-    of the requested table in a single call).
-
-    Total BEA API usage per live run is one request per indicator
-    (currently 2 — GDP + Personal Income), well under the 1000-per-day
-    free-tier budget. The registry includes entries with
-    ``api_fetch=False`` (GDP, schedule-only) — the probe still exercises
-    their API shape so an operator can verify the parser's expectations
-    before enabling an API lane.
-    """
-    now_year = datetime.now(timezone.utc).year
-    probes: list[BEAProbe] = []
-    for series_id, spec in BEA_INDICATOR_REGISTRY.items():
-        token = _bea_probe_token(spec.indicator)
-        probes.append(
-            BEAProbe(
-                name=f"{token}_two_year_window",
-                series_id=series_id,
-                indicator=spec.indicator,
-                dataset=spec.dataset,
-                table=spec.table,
-                line_number=spec.line_number,
-                frequency=spec.frequency,
-                description=spec.title,
-                start_year=now_year - 1,
-                end_year=now_year,
-            )
-        )
-    return probes
-
-
-def _bea_probe_token(indicator_label: str) -> str:
-    """Slugify an indicator label into a stable probe-name token."""
-    token = indicator_label.strip().lower()
-    token = re.sub(r"[^a-z0-9]+", "_", token).strip("_")
-    return token or "indicator"
-
-
-def _diff_bea_observation(raw_row: dict[str, Any]) -> RowDiff:
-    """Field diff for one BEA observation row (the ``raw`` dict).
-
-    Parser reads ``TimePeriod`` (→ reference date) and ``DataValue``
-    (→ numeric); hashes on ``DataValue`` + ``LineDescription`` +
-    ``NoteRef`` (methodology / release-stage flags). ``LineNumber`` is
-    the identity coordinate on the line within the table. Any other
-    field upstream returns — ``SeriesCode``, ``CL_UNIT``,
-    ``UNIT_MULT``, ``METRIC_NAME`` are documented possibilities on some
-    BEA datasets — surfaces as UNKNOWN_OBSERVED so the operator decides
-    whether to read or ignore it.
-    """
-    observed = set(raw_row.keys())
-    diff = RowDiff(
-        observed_fields=sorted(observed),
-        read_by_parser=sorted(observed & BEA_OBS_EXPECTED_FIELDS),
-        ignored_by_parser=sorted(
-            (BEA_OBS_EXPECTED_FIELDS & observed) - BEA_OBS_EXPECTED_FIELDS
-        ),  # empty by construction — symmetry with BLS shape
-        unknown_observed=sorted(observed - BEA_OBS_EXPECTED_FIELDS),
-        missing_expected=sorted(BEA_OBS_EXPECTED_FIELDS - observed),
-    )
-
-    time_period = raw_row.get("TimePeriod")
-    if time_period is None or str(time_period).strip() == "":
-        diff.type_warnings.append(
-            "TimePeriod missing — client would synthesize an empty reference date"
-        )
-    data_value = raw_row.get("DataValue")
-    if data_value is None:
-        diff.type_warnings.append("DataValue missing — parser would write null actual")
-    elif not isinstance(data_value, str):
-        # BEA documents DataValue as a string ("21,542.5" / "(NA)" /
-        # "(D)") so the client's ``_parse_data_value`` can run its
-        # suppressed-value detection. A non-string would bypass the
-        # suppressed-value check and land as ``None``.
-        diff.type_warnings.append(
-            f"DataValue is {type(data_value).__name__}={data_value!r} — "
-            f"parser expects str"
-        )
-    line_number = raw_row.get("LineNumber")
-    if line_number is None:
-        diff.type_warnings.append(
-            "LineNumber missing — line filter + series-id synthesis both break"
-        )
-    return diff
-
-
-def _try_parse_bea(obs) -> tuple[bool, str]:
-    """Dry-run the calendar-side BEA projector on one observation."""
-    spec = BEA_INDICATOR_REGISTRY.get(obs.series_id)
-    if spec is None:
-        return False, f"no BEA calendar spec for series_id={obs.series_id!r}"
-    try:
-        raw, event = parse_bea_observation(
-            obs, snapshot_epoch_ms=1_700_000_000_000, spec=spec,
-        )
-        return True, (
-            f"ok indicator={spec.indicator} event_id={raw.provider_event_id[:10]}…"
-        )
-    except Exception as exc:
-        return False, f"{type(exc).__name__}: {exc}"
-
-
-def run_bea_probe(client: BEAClient, probe: BEAProbe) -> ProbeResult:
-    """Execute one BEA probe and return a populated :class:`ProbeResult`.
-
-    Runs via ``BEAClient.get_data`` — reuses the production transport so
-    throttling + error handling behave identically to a recurring fetch.
-    BEA returns every line for the requested table; the runner filters
-    down to the probe's ``line_number`` after the call lands. An empty
-    filtered set lands as ``ok`` with ``row_count=0`` and a note that
-    records the full-table row count — the signal that the coordinate
-    drifted upstream vs the registry.
-    """
-    generic = Probe(
-        name=probe.name,
-        path=(
-            f"GET https://apps.bea.gov/api/data?DatasetName={probe.dataset}"
-            f"&TableName={probe.table}"
-        ),
-        description=probe.description,
-        expected_shape="list[BEAObservation] after client parse",
-        expected_fields=BEA_OBS_EXPECTED_FIELDS,
-    )
-    result = ProbeResult(probe=generic, status="skipped")
-    year_param = ",".join(
-        str(y) for y in range(probe.start_year, probe.end_year + 1)
-    )
-    result.request_path = (
-        f"{generic.path}"
-        f"&Frequency={probe.frequency}&Year={year_param} "
-        f"(filter LineNumber={probe.line_number})"
-    )
-
-    if not client.api_key:
-        result.status = "auth_missing"
-        result.notes.append("BEA_API_KEY not set — probe skipped")
-        return result
-
-    import time as _time
-    t0 = _time.monotonic()
-    try:
-        observations = client.get_data(
-            probe.dataset,
-            TableName=probe.table,
-            Frequency=probe.frequency,
-            Year=year_param,
-        )
-    except Exception as exc:
-        result.status = "http_error"
-        result.notes.append(f"{type(exc).__name__}: {exc}")
-        return result
-    result.http_elapsed_ms = (_time.monotonic() - t0) * 1000
-
-    # BEA returns every line of the table; filter down to the probe's
-    # line before shape-diffing so the sample row reflects the indicator
-    # we actually care about rather than an arbitrary neighbour line.
-    filtered = [o for o in observations if o.line_number == probe.line_number]
-
-    result.status = "ok"
-    result.row_count = len(filtered)
-    if not filtered:
-        result.notes.append(
-            f"BEA returned zero observations for LineNumber={probe.line_number} "
-            f"(table payload had {len(observations)} rows across other lines) "
-            f"— coordinate may have drifted upstream"
-        )
-        return result
-
-    # BEA ``obs.date`` is end-of-period ISO (quarterly/annual) or first-
-    # of-month (monthly) — lexical sort descending puts the newest
-    # observation first for both.
-    filtered = sorted(filtered, key=lambda o: o.date, reverse=True)
-    sample = filtered[0]
-    result.sample_row = sample.raw or {
-        "series_id": sample.series_id,
-        "date": sample.date,
-        "value": sample.value,
-        "line_number": sample.line_number,
-    }
-    if sample.raw:
-        result.field_diff = _diff_bea_observation(sample.raw)
-
-    # TimePeriod tally — surfaces mixed-frequency responses (e.g. an
-    # annual row landing inside a quarterly ``Frequency=Q`` query) that
-    # would otherwise silently inflate the event stream.
-    period_counter: Counter = Counter()
-    for obs in filtered:
-        period_counter[repr(obs.raw.get("TimePeriod") if obs.raw else None)] += 1
-    result.enum_counters = {"TimePeriod": period_counter}
-
-    sample_n = min(10, len(filtered))
-    result.parse_attempts = sample_n
-    for obs in filtered[:sample_n]:
-        ok, msg = _try_parse_bea(obs)
-        if ok:
-            result.parse_successes += 1
-        else:
-            if len(result.parse_error_samples) < 3:
-                result.parse_error_samples.append(msg)
-    return result
-
-
-# ──────────────────────────────────────────────────────────────────────────
-# Census probe plan + runner
-# ──────────────────────────────────────────────────────────────────────────
-
-CENSUS_EITS_EXPECTED_FIELDS: frozenset[str] = frozenset({
-    "data_type_code",
-    "seasonally_adj",
-    "category_code",
-    "cell_value",
-    "error_data",
-    "time_slot_id",
-    "time_slot_name",
-    "time",
-    "us",
-})
-
-
-@dataclass
-class CensusProbe:
-    """One Census live-validation probe — one series in one EITS year."""
-
-    name: str
-    series_id: str
-    dataset: str
-    data_type_code: str
-    category_code: str
-    seasonally_adj: str
-    time_slot_id: str
-    description: str
-    year: int
-
-
-def plan_census_probes() -> list[CensusProbe]:
-    """One probe per Census calendar registry entry."""
-    now_year = datetime.now(timezone.utc).year
-    probes: list[CensusProbe] = []
-    for series_id, spec in CENSUS_INDICATOR_REGISTRY.items():
-        token = _bls_probe_token(spec.indicator)
-        probes.append(
-            CensusProbe(
-                name=f"{token}_{now_year}",
-                series_id=series_id,
-                dataset=spec.dataset,
-                data_type_code=spec.data_type_code,
-                category_code=spec.category_code,
-                seasonally_adj=spec.seasonally_adj,
-                time_slot_id=spec.time_slot_id,
-                description=spec.title,
-                year=now_year,
-            )
-        )
-    return probes
-
-
-def _diff_census_row(raw_row: dict[str, Any]) -> RowDiff:
-    observed = set(raw_row.keys())
-    diff = RowDiff(
-        observed_fields=sorted(observed),
-        read_by_parser=sorted(observed & CENSUS_EITS_EXPECTED_FIELDS),
-        ignored_by_parser=[],
-        unknown_observed=sorted(observed - CENSUS_EITS_EXPECTED_FIELDS),
-        missing_expected=sorted(CENSUS_EITS_EXPECTED_FIELDS - observed),
-    )
-    value = raw_row.get("cell_value")
-    if value in (None, ""):
-        diff.type_warnings.append("cell_value missing")
-    else:
-        try:
-            float(str(value))
-        except ValueError:
-            diff.type_warnings.append(f"cell_value is not numeric: {value!r}")
-    return diff
-
-
-def _row_matches_census_probe(row: dict[str, str], probe: CensusProbe) -> bool:
-    return (
-        row.get("data_type_code") == probe.data_type_code
-        and row.get("seasonally_adj") == probe.seasonally_adj
-        and row.get("category_code") == probe.category_code
-        and row.get("time_slot_id") == probe.time_slot_id
-    )
-
-
-def _census_obs_from_row(row: dict[str, str], probe: CensusProbe) -> CensusEITSObservation:
-    return CensusEITSObservation(
-        series_id=probe.series_id,
-        dataset=probe.dataset,
-        time=row.get("time", ""),
-        data_type_code=row.get("data_type_code", ""),
-        category_code=row.get("category_code", ""),
-        seasonally_adj=row.get("seasonally_adj", ""),
-        time_slot_id=row.get("time_slot_id", ""),
-        time_slot_name=row.get("time_slot_name", ""),
-        cell_value=row.get("cell_value", ""),
-        error_data=row.get("error_data", ""),
-        raw=dict(row),
-    )
-
-
-def _try_parse_census(row: dict[str, str], probe: CensusProbe) -> tuple[bool, str]:
-    spec = CENSUS_INDICATOR_REGISTRY.get(probe.series_id)
-    if spec is None:
-        return False, f"no Census calendar spec for series_id={probe.series_id!r}"
-    try:
-        obs = _census_obs_from_row(row, probe)
-        raw, event = parse_census_observation(
-            obs,
-            snapshot_epoch_ms=1_700_000_000_000,
-            spec=spec,
-        )
-        return True, (
-            f"ok indicator={spec.indicator} event_id={raw.provider_event_id[:10]}…"
-        )
-    except Exception as exc:
-        return False, f"{type(exc).__name__}: {exc}"
-
-
-def run_census_probe(client: CensusEITSClient, probe: CensusProbe) -> ProbeResult:
-    generic = Probe(
-        name=probe.name,
-        path=f"GET https://api.census.gov/data/timeseries/eits/{probe.dataset}",
-        description=probe.description,
-        expected_shape="list[dict] after EITS header-row decode",
-        expected_fields=CENSUS_EITS_EXPECTED_FIELDS,
-    )
-    result = ProbeResult(probe=generic, status="skipped")
-    result.request_path = (
-        f"{generic.path}?time={probe.year}&for=us:* "
-        f"(filter data_type_code={probe.data_type_code} "
-        f"seasonally_adj={probe.seasonally_adj} "
-        f"category_code={probe.category_code})"
-    )
-
-    import time as _time
-    t0 = _time.monotonic()
-    try:
-        rows = client.get_dataset_year(probe.dataset, probe.year)
-    except Exception as exc:
-        result.status = "http_error"
-        result.notes.append(f"{type(exc).__name__}: {exc}")
-        return result
-    result.http_elapsed_ms = (_time.monotonic() - t0) * 1000
-
-    filtered = [row for row in rows if _row_matches_census_probe(row, probe)]
-    result.status = "ok"
-    result.row_count = len(filtered)
-    if not filtered:
-        result.notes.append(
-            f"Census returned zero rows for {probe.series_id} in {probe.year} "
-            f"(dataset payload had {len(rows)} rows)"
-        )
-        return result
-
-    filtered = sorted(filtered, key=lambda row: row.get("time", ""), reverse=True)
-    result.sample_row = filtered[0]
-    result.field_diff = _diff_census_row(filtered[0])
-    result.enum_counters = {
-        "time": Counter(row.get("time") for row in filtered),
-        "error_data": Counter(row.get("error_data") for row in filtered),
-    }
-
-    sample_n = min(10, len(filtered))
-    result.parse_attempts = sample_n
-    for row in filtered[:sample_n]:
-        ok, msg = _try_parse_census(row, probe)
-        if ok:
-            result.parse_successes += 1
-        else:
-            if len(result.parse_error_samples) < 3:
-                result.parse_error_samples.append(msg)
-    return result
 
 
 # ──────────────────────────────────────────────────────────────────────────
